+++ finagle-netty3/src/main/scala/com/twitter/finagle/netty3/transport/ChannelTransport.scala
@@ -5,7 +5,6 @@ import com.twitter.finagle.transport.Transport
 import com.twitter.finagle.{ChannelClosedException, ChannelException, Status}
 import com.twitter.util.{Future, Promise, Return, Time}
 import java.net.SocketAddress
-import java.security.cert.Certificate
 import java.util.concurrent.atomic.AtomicBoolean
 import org.jboss.netty.channel._
 
@@ -157,7 +156,6 @@ class ChannelTransport[In, Out](ch: Channel)
 
   def localAddress: SocketAddress = context.localAddress
   def remoteAddress: SocketAddress = context.remoteAddress
-  def peerCertificate: Option[Certificate] = context.peerCertificate
 
   override def toString = s"Transport<channel=$ch, onClose=$closep>"